icone_identidade v0.1.0 IconeIdentidade

Documentation for IconeIdentidade.

Functions

constructor_pixel(image)

Receive an 'Imagem' anda based on ':grid' value, generate start and final pixels position

##Exemple
  iex> IconeIdentidade.constructor_pixel(%IconeIdentidade.Imagem{color: nil,grid: [{46, 1},{176, 2},{46, 3}, {98, 5},{236, 6},{236, 8}, {98, 9},],hex: nil,pixel_map: nil})
  %IconeIdentidade.Imagem{
    color: nil,
    grid: [{46, 1}, {176, 2}, {46, 3}, {98, 5}, {236, 6}, {236, 8}, {98, 9}],
    hex: nil,
    pixel_map: [
      {{50, 0}, {100, 50}},
      {{100, 0}, {150, 50}},
      {{150, 0}, {200, 50}},
      {{0, 50}, {50, 100}},
      {{50, 50}, {100, 100}},
      {{150, 50}, {200, 100}},
      {{200, 50}, {250, 100}}
    ]
  }

create_color(image)

create_color(IconeIdentidade.Imagem.t()) :: IconeIdentidade.Imagem.t()

Create color who will fill generated image Use to this, 3 first numbers generated of :grid list, use :color to record this

##Exemple
  iex> IconeIdentidade.create_color(%IconeIdentidade.Imagem{color: nil, grid: nil, hex: [161, 46, 176, 98, 236, 169, 209, 230, 198, 159, 207, 139, 96, 55, 135,195], pixel_map: nil})
  %IconeIdentidade.Imagem{
    color: {161, 46, 176},
    grid: nil,
    hex: [161, 46, 176, 98, 236, 169, 209, 230, 198, 159, 207, 139, 96, 55, 135,
    195],
    pixel_map: nil
  }

create_table(image)

'create_table' receive an '%IconeIdentidade.Imagem' and create a grid, where will be maped the print screen

##Exemple
  iex> IconeIdentidade.create_table(%IconeIdentidade.Imagem{color: {161, 46, 176}, grid: nil, hex: [161, 46, 176, 98, 236, 169, 209, 230, 198, 159, 207, 139, 96, 55, 135, 195], pixel_map: nil})
  %IconeIdentidade.Imagem{
    color: {161, 46, 176},
    grid: [
      {161, 0},
      {46, 1},
      {176, 2},
      {46, 3},
      {161, 4},
      {98, 5},
      {236, 6},
      {169, 7},
      {236, 8},
      {98, 9},
      {209, 10},
      {230, 11},
      {198, 12},
      {230, 13},
      {209, 14},
      {159, 15},
      {207, 16},
      {139, 17},
      {207, 18},
      {159, 19},
      {96, 20},
      {55, 21},
      {135, 22},
      {55, 23},
      {96, 24}
    ],
    hex: [161, 46, 176, 98, 236, 169, 209, 230, 198, 159, 207, 139, 96, 55, 135,
    195],
    pixel_map: nil
  }

draw_pixel(imagem)

Generate image joining all previous information

##Exemple

hash_input(input)

Function receve text and past Struct 'IconeIdentidade.Imagem' with text buffered at atom :hex in a list after have been md5 coded

##Exemple
  iex> IconeIdentidade.hash_input("Elixir")
  %IconeIdentidade.Imagem{
    color: nil,
    grid: nil,
    hex: [161, 46, 176, 98, 236, 169, 209, 230, 198, 159, 207, 139, 96, 55, 135,
    195],
    pixel_map: nil
  }

main(input)

Main function, insert any string to generate a pixel_map refering the text

Exemple

iex> IconeIdentidade.main("Elixir")
:ok

odd_remove(image)

Function used to remove grids who has odd value

##Exemple
  iex> IconeIdentidade.odd_remove(%IconeIdentidade.Imagem{ color: {161, 46, 176}, grid: [ {161, 0}, {46, 1}, {176, 2}, {46, 3}, {161, 4}, {98, 5}, {236, 6}, {169, 7}, {236, 8}, {98, 9}, {209, 10}, {230, 11}, {198, 12}, {230, 13}, {209, 14}, {159, 15}, {207, 16}, {139, 17}, {207, 18}, {159, 19}, {96, 20}, {55, 21}, {135, 22}, {55, 23}, {96, 24} ], hex: [161, 46, 176, 98, 236, 169, 209, 230, 198, 159, 207, 139, 96, 55, 135, 195], pixel_map: nil })
  %IconeIdentidade.Imagem{
    color: {161, 46, 176},
    grid: [
      {46, 1},
      {176, 2},
      {46, 3},
      {98, 5},
      {236, 6},
      {236, 8},
      {98, 9},
      {230, 11},
      {198, 12},
      {230, 13},
      {96, 20},
      {96, 24}
    ],
    hex: [161, 46, 176, 98, 236, 169, 209, 230, 198, 159, 207, 139, 96, 55, 135,
    195],
    pixel_map: nil
  }

save_image(image, input_inicial)

Save the file in project directory using the initial input as name.

shorter(phrase)

Print the firsts letters in each word in the phrase

   ##Expemple
    iex> IconeIdentidade.shorter("your won phrase with many words")
    "YWPWMW"

to_mirror(linha)

To mirror two first position in a list to the lestest position in the same list

##Exemple

iex> IconeIdentidade.to_mirror([1,2,3])
[1,2,3,2,1]

iex> IconeIdentidade.to_mirror([Pedro, Paulo, Carlos])
[Pedro, Paulo, Carlos, Paulo, Pedro]